Move json,yaml,jsonpath printers to correct locations
